Moving gcp scripts inside utils
