+++ bqsqoop/utils/gcloud/auth.py
