+++ bqsqoop/utils/gcloud/storage.py
